pyproject.toml
[project]
name = "python-ai-agent"
version = "0.1.0"
description = "Building a basic AI agent in Python using the Gemini API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
    "rich>=14.3.3",
]
